[Refactor](Keymappings): Moved diff mappings to separate file.
